+++ earth2studio/models/px/__init__.py
@@ -38,6 +38,7 @@ from earth2studio.models.px.fcn import FCN
 from earth2studio.models.px.fcn3 import FCN3
 from earth2studio.models.px.fengwu import FengWu
 from earth2studio.models.px.fuxi import FuXi
+from earth2studio.models.px.fuxi_s2s import FuXiS2S
 from earth2studio.models.px.gencast_mini import GenCastMini
 from earth2studio.models.px.graphcast_operational import GraphCastOperational
 from earth2studio.models.px.graphcast_small import GraphCastSmall